hw2/api_gateway/requirements.txt
fastapi
uvicorn
httpx
python-jose[cryptography]
pydantic-settings
